Use same opts/handler for exec mixin wiring functions
