+++ pkg/exec/testdata/install_input.yaml
