+++ pkg/exec/testdata/upgrade_input.yaml
@@ -1,0 +1,7 @@
+upgrade:
+- exec:
+    description: "Install Hello World"
+    command: bash
+    arguments:
+    - -c
+    - echo Hello World
